+++ lib/src/presentation/bloc/analisis/hn/analisis_represtamo/analisis_represtamo_state.dart
@@ -28,7 +28,7 @@ class AnalisisReprestamoState extends Equatable {
   final int totalDisponibleActivo;
   final int adelantoProveedores;
   final int incobrables;
-  final int totalCuentasXCobrar;
+  final double totalCuentasXCobrar;
   final double totalInventario;
   final int totalActivosFijos;
   final double totalActivos;
@@ -408,7 +408,7 @@ class AnalisisReprestamoState extends Equatable {
     int? totalDisponibleActivo,
     int? adelantoProveedores,
     int? incobrables,
-    int? totalCuentasXCobrar,
+    double? totalCuentasXCobrar,
     double? totalInventario,
     int? totalActivosFijos,
     double? totalActivos,
